Add a post-merge hook that enforces our git workflow.
This checks a whole pile of attributes of commits, looking for things that
commonly go wrong or violate our standards, and emits appropriate warnings.

It also takes the liberty: if the merge isn't right, it rejects it out of
hand.  Hard reset back to the previous state, bam.  No trusting you to do it
here, just outright cruelty and process enforcement.